Fix resource loading

+++ cli/src/main/resources/install-linux.sh
@@ -1,14 +1,12 @@
 #!/bin/bash
 
-trap 'kill -TERM $PID' TERM INT
-
 ST_INSTALL_LOC=$ST_INSTALL_LOC
 
-"${ST_INSTALL_LOC}"jre/bin/java -classpath "${ST_INSTALL_LOC}cli/*" io.supertokens.cli.Main false "${ST_INSTALL_LOC}" $@ &
-PID=$!
-wait $PID
-if [ $? -eq 0 ] && [ "$#" -ne 0 ] && [ "$1" == update ]; then
-  "${ST_INSTALL_LOC}"/.update/supertokensExe update-complete --originalInstallDir="${ST_INSTALL_LOC}" &
+if grep docker /proc/1/cgroup -qa; then
+  trap 'kill -TERM $PID' TERM INT
+  "${ST_INSTALL_LOC}"jre/bin/java -classpath "${ST_INSTALL_LOC}cli/*" io.supertokens.cli.Main false "${ST_INSTALL_LOC}" $@ &
   PID=$!
   wait $PID
+else
+  "${ST_INSTALL_LOC}"jre/bin/java -classpath "${ST_INSTALL_LOC}cli/*" io.supertokens.cli.Main false "${ST_INSTALL_LOC}" $@
 fi
